+++ Include/test.hpp
@@ -1,0 +1,17 @@
+//
+// Created by Matthis Doquocbao on 4/3/23.
+//
+
+#ifndef WEBSREVER_TEST_HPP
+#define WEBSREVER_TEST_HPP
+
+
+class test {
+    int i;
+    int d;
+
+
+};
+
+
+#endif //WEBSREVER_TEST_HPP
