+++ Unit_test/Sources/UTmain.cpp
@@ -11,8 +11,9 @@
 /* ************************************************************************** */
 
 #include "../Include/unit_test.hpp"
+#include "gtest/gtest.h"
 
-int main() {
+int TEST() {
     std::cout <<"====================[ Unit test ]===============================\n" <<std::endl;
     enum{constructor,}; // add enum for name of test
     int nbr = 1;//add number for each function
@@ -28,5 +29,10 @@ int main() {
     }
     std::cout <<"\n================[ Unit_test | Success ]=========================" <<std::endl;
     exit(EXIT_SUCCESS);
+}
+
+int main(int argc, char** argv){
+    ::testing::InitGoogleTest(&argc, &argv);
+    return RUN_ALL_TESTS();
 }
