+++ usr->local/var/www/webserv1.com/css/style.css
@@ -6,75 +6,148 @@ html {
     box-sizing: inherit;
 }
 
-#console {
+footer {
+    background-color: #5e5c66;
+    color: white;
+    text-align: center;
+    padding: 0.5em;
+}
+
+header p{
+    background-color: #5e5c66;
+    color: white;
+    text-align: center;
+    font-size: 1.5em;
+    padding: 0.5em;
+}
+
+#terminal {
     display: flex;
     flex-direction: column;
-    height: 30em;
-    border: 1px solid #0aa;
-    font-family: sans-serif;
-    word-wrap: break-word;
+    height: 35em;
+    border: 0.2em solid gray;
+
 }
 
-#console .header footer{
+#terminal .write {
     display: flex;
-    align-items: center;
-    height: 2em;
-    padding: 0 0 0 .5em;
-}
-
-#console .main {
-    background-color: #000;
-    color: #ddd;
+    background-color: black;
     height: 100%;
-    overflow-y: scroll;
-    font-family: monospace;
-    font-size: 1.1em;
-    line-height: 1.1em;
+    width: 100%;
+    word-wrap: break-word;
+    /*color: #ddd;*/
+    /*height: 100%;*/
+    /*overflow-y: scroll;*/
+    /*font-family: monospace;*/
+    /*font-size: 1.1em;*/
+    /*line-height: 1.1em;*/
 }
 
-nav {
-    display: flex;
-    align-items: stretch;
-    background-color: #302e36;
+#terminal .head{
+    background-color: #222222;
+    width: 100%;
+    height: 2em;
+    color: white;
+    text-align: center;
+    align-content: center;
 }
 
-nav a {
-    margin: 0 1px 0 0;
+#terminal .close{
+    position: absolute;
+    background-color: firebrick;
+    height: 1.5em;
+    width: 1.5em;
+    text-align: center;
+    padding: 0.1em;
+    border: 0.2em solid darkred;
 }
 
-nav a:hover {
-    background-color: #4d4b53;
+#terminal .close:hover{
+    background-color: indianred;
 }
 
-nav a.active {
-    background-color: #5e5c66;
+#terminal .close:active{
+    background-color: crimson;
 }
 
-nav a.active:hover {
-    background-color: #777481;
-}
+/*#terminal .header footer{*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    height: 2em;*/
+/*    padding: 0 0 0 .5em;*/
+/*}*/
 
-nav a > div {
-    padding: .3em .3em;
-    margin: 0 .2em;
-    border-bottom: 3px solid#0aa;
-    transition: all .1s ease;
-}
 
-nav a.active > div {
-    margin: 0;
-    padding: .3em .5em;
-}
 
-nav a:hover > div {
-    margin: 0 0;
-    padding: .3em .5em;
-}
+/*nav {*/
+/*    display: flex;*/
+/*    align-items: stretch;*/
+/*    background-color: #302e36;*/
+/*}*/
 
-nav img {
-    width: 32px;
-    height: 32px;
-    background: linear-gradient(to bottom, transparent 3px, #fff 3px, #fff 26px, transparent 0 );
-    padding: 0 1px;
-    border-radius: 5px;
-}
+/*nav a {*/
+/*    margin: 0 1px 0 0;*/
+/*}*/
+
+/*nav a:hover {*/
+/*    background-color: #4d4b53;*/
+/*}*/
+
+/*nav a.active {*/
+/*    background-color: #5e5c66;*/
+/*}*/
+
+/*nav a.active:hover {*/
+/*    background-color: #777481;*/
+/*}*/
+
+/*nav a > div {*/
+/*    padding: .3em .3em;*/
+/*    margin: 0 .2em;*/
+/*    border-bottom: 3px solid#0aa;*/
+/*    transition: all .1s ease;*/
+/*}*/
+
+/*nav a.active > div {*/
+/*    margin: 0;*/
+/*    padding: .3em .5em;*/
+/*}*/
+
+/*nav a:hover > div {*/
+/*    margin: 0 0;*/
+/*    padding: .3em .5em;*/
+/*}*/
+
+/*nav img {*/
+/*    width: 32px;*/
+/*    height: 32px;*/
+/*    background: linear-gradient(to bottom, transparent 3px, #fff 3px, #fff 26px, transparent 0 );*/
+/*    padding: 0 1px;*/
+/*    border-radius: 5px;*/
+/*}*/
+
+/*#console {*/
+/*    display: flex;*/
+/*    flex-direction: column;*/
+/*    height: 30em;*/
+/*    border: 1px solid #0aa;*/
+/*    font-family: sans-serif;*/
+/*    word-wrap: break-word;*/
+/*}*/
+
+/*#console .header footer{*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    height: 2em;*/
+/*    padding: 0 0 0 .5em;*/
+/*}*/
+
+/*#console .main {*/
+/*    background-color: #000;*/
+/*    color: #ddd;*/
+/*    height: 100%;*/
+/*    overflow-y: scroll;*/
+/*    font-family: monospace;*/
+/*    font-size: 1.1em;*/
+/*    line-height: 1.1em;*/
+/*}*/
