+++ usr->local/var/www/webserv1.com/terminal.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Terminal</title>
+
+    <link rel="stylesheet" href="./css/reset.css">
+    <link rel="stylesheet" href="./css/style.css">
+    <script src="./js/terminal.js"></script>
+</head>
+<body>
+    <header>
+        <p>Webwerv</p>
+    </header>
+    <div id="terminal">
+        <div class="close">X</div>
+        <div class="head">Terminal</div>
+        <div class="write"></div>
+    </div>
+    <footer>
+        <p>&copy; Copyright Matthis DoQuocBao | 42 school </p>
+    </footer>
+
+</body>
+</html>
